requirements-dev.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
dotenv==0.9.9
et_xmlfile==2.0.0
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
multidict==6.5.0
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
